level-three/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
urllib3>=2.2.0
pandas>=2.2.0
python-docx>=1.1.0
reportlab>=4.1.0
elasticsearch>=8.12.0
